+++ lib/yt/channel.rb
@@ -91,6 +91,11 @@ module Yt
     #   channels module.
     has_attribute :featured_channels_urls, in: %i(branding_settings channel), default: []
 
+    # @!attribute [r] related_playlists
+    # @return [Hash] the playlists associated with the channel, such as the
+    #   channel's uploaded videos, liked videos, and watch history.
+    has_attribute :related_playlists, in: :content_details
+
     # @return [String] the canonical form of the channel’s URL.
     def canonical_url
       "https://www.youtube.com/channel/#{id}"
@@ -134,12 +139,27 @@ module Yt
       end
     end
 
+    # @return [Yt::Relation<Yt::Playlist>] the playlists associated with
+    #   liked videos. Includes the deprecated favorites if still present.
+    def like_playlists
+      @like_lists ||= Relation.new(Playlist, ids: like_list_ids) do |options|
+        get '/youtube/v3/playlists', resource_params(options)
+      end
+    end
+
     # @return [Yt::Channel] the channel associated with the YouTube account
     #   that provided the authentication token.
     def self.mine
       Relation.new(self) do |options|
         get '/youtube/v3/channels', mine: true, part: 'id'
       end.first
+    end
+
+  private
+
+    def like_list_ids
+      names = %w(likes favorites)
+      related_playlists.select{|name,_| names.include? name}.values
     end
   end
 end